# kmarket-backend/docs/openapi-admins.yaml
openapi: 3.0.3
info:
  title: Admin Management API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
paths:
  /admins:
    get:
      summary: List admin accounts
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  page: { type: integer }
                  pageSize: { type: integer }
                  total: { type: integer }
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUser'
        '401': { description: Unauthorized }
    post:
      summary: Create admin account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                walletAddress: { type: string }
                permissions: { type: object }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
        '400': { description: Bad Request }
        '401': { description: Unauthorized }
  /admins/{id}:
    get:
      summary: Get admin by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '404': { description: Not Found }
    put:
      summary: Update admin
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
                permissions: { type: object }
                status: { type: string, enum: [active, disabled] }
      responses:
        '204': { description: No Content }
        '401': { description: Unauthorized }
    delete:
      summary: Soft delete admin
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: No Content }
        '401': { description: Unauthorized }

components:
  schemas:
    AdminUser:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        username: { type: string }
        permissions: { type: object }
        status: { type: string }
        address: { type: string }
        role: { type: string }
        createdAt: { type: integer, description: 'timestamp ms' }
        updatedAt: { type: integer, description: 'timestamp ms' }